% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSim_dev.R
\name{mort}
\alias{mort}
\title{mort(): sets death year to a constant for some members of the population.}
\usage{
mort(indiv = makeFounders(), year = "-1", type = "simple",
  maxAge = Inf, maxPop = 1000, mortRate, ageMort, stockMort,
  ageStockMort)
}
\arguments{
\item{indiv}{A matrix of inidviduals, as from makeFounders(), move(), or mate().}

\item{type}{One of "simple", "flat", "age", "stock", or "ageStock".
If type = "simple" and the living pop > maxPop, individuals are killed at random until
the living pop == maxPop. Can easily be set to never cause extinctions.
If type = "flat", individuals are killed with probability set in mortRate. Generates
an exponential death curve.
If type = "age", individuals are killed with probability for their age set in ageMort.
If type = "stock", individuals are killed with probability for their stock set in
stockMort.
If type = "ageStock", individuals are killed with probability for their age:stock
combination set in ageStockMort.}

\item{maxAge}{Sets an age at which animals *will* be killed before anything else happens. Allows
a short age-specific mortality curve to be set, without checking if there are any
individuals outside the range for each iteration.}

\item{maxPop}{If type = "simple", the population will be reduced to this number, if not already
smaller. See 'type'.}

\item{mortRate}{Numeric mortality rate. See 'type'.}

\item{ageMort}{Numeric vector of mortality rates, one for each age, ordered 0:max(age). See 'type'.}

\item{stockMort}{Numeric vector of mortality rates, one for each stock, ordered 1:max(stock). Note
that stocks are numbered (as in makeFounders() ), not named. Because stocks are
stored as a character vector, stocks are converted via as.numeric() to associate
rates with stocks. This distinction is important in cases with >9 stocks. See
'type'.}

\item{ageStockMort}{A matrix of mortality rates, with age by row and stock by column. See 'ageMort'
and 'stockMort' for structure of rows and columns.}
}
\description{
Members are chosen according to one of several defined mortality structures.
Mortality rates can bring the population to a specified size, or can be a flat probability,
or a probability that depends on age, stock, or age:stock.
}
